Initial Commit
This commit defines the working space that this project will be using
— the "git repository".

By convention, there is very little that's part of the initial commit.
In this way, when we build things in the future we can add them one
chunk at a time, documenting in version control (git¹).

In this commit, the repository was prepared by creating a new, fresh
directory called "resume". Within that directory, running the
commands in the terminal:

  $ git init --initial-branch=main

Creating the repository. To create the resume, the command:

  $ cat <<'EOF' > README.md
  # Resume
  ... (abbreviated)
  EOF

Was used to generate the README.md file. The file is then added to
the repository:

  $ git add README.md

And "committed" to version control:

  $ git commit --message="Initial Commit"

== Design Notes
=== Tutorial

This project is designed to be the project which people can use as a
reference for the "practical introduction to software development"
course. Given this, it should clearly explain each of it steps, for
someone with a minimal background in software development.

In practicality, writing a course is difficult, but writing software
is fun — this serves also as a mechanism to chip away at the courses,
but to also keep motivated.

=== Git Message

Strictly speaking, the message was written while writing the code
itself, and saved to a file called ".COMMIT". The command to commit
it to version control was then:

  $ git commit --file .COMMIT

== Links
[¹]: Git SCM. (2022). git. https://h4n.link/uEsh